+++ .github/workflows/pypi-wheels.yml
@@ -61,7 +61,7 @@ jobs:
          -DCMAKE_LINKER:FILEPATH="C:/Program Files/LLVM/bin/lld-link.exe"
 
       - name: "Build Project"
-        run: cmake --build build --target asset pypiwheel --config Release -- -j1
+        run: cmake --build build --target asset pypiwheel --config Release -- -j2
 
       - name: 'Upload Wheels'
         uses: actions/upload-artifact@v3
@@ -105,7 +105,7 @@ jobs:
             run: cmake -S . -B build -G "Unix Makefiles" -DCMAKE_BUILD_TYPE=Release -DPIP_INSTALL=True -DBUILD_ASSET_WHEEL=True
 
           - name: "Build Project"
-            run: cmake --build build --target asset pypiwheel --config Release -- -j1
+            run: cmake --build build --target asset pypiwheel --config Release -- -j2
 
           - name: Audit Wheel Test
             run:  auditwheel show build/pypiwheel/asset_asrl/dist/*.whl
@@ -119,55 +119,7 @@ jobs:
                 name: dist_${{ matrix.python-version }}_${{ matrix.platform }}
                 path: wheelhouse/*.whl
                 retention-days: 5
- build_linux_compat:
-        strategy:
-              fail-fast: false
-              matrix:
-                platform: [ubuntu-20.04]
-                python-version: ["3.7","3.8","3.9","3.10","3.11"]
-                compiler: ["clang-12"]
-
-        runs-on: ${{ matrix.platform }}
-
-        steps:
-          - uses: actions/checkout@v3
-            with:
-                submodules: true
-
-
-          - uses: actions/setup-python@v3
-            with:
-              python-version: ${{ matrix.python-version }}
-
-          - name: Install Ubuntu dependencies
-            env:
-                COMPILER: ${{ matrix.compiler }}
-            run: bash $GITHUB_WORKSPACE/.github/install_ubuntu_dependencies.sh
-
-          - name: Add python requirements
-            run: python -m pip install --upgrade wheel setuptools && python -m pip install -r requirements.txt
-
-          - run: pip install patchelf auditwheel
-
-
-          - name: "Create build directory and run CMake"
-            run: cmake -S . -B build -G "Unix Makefiles" -DCMAKE_BUILD_TYPE=Release -DPIP_INSTALL=True -DBUILD_ASSET_WHEEL=True
-
-          - name: "Build Project"
-            run: cmake --build build --target asset pypiwheel --config Release -- -j1
-
-          - name: Audit Wheel Test
-            run:  auditwheel show build/pypiwheel/asset_asrl/dist/*.whl
-
-          - name: Audit Wheel Repair
-            run:  auditwheel repair --plat manylinux_2_31_x86_64 build/pypiwheel/asset_asrl/dist/*.whl
-
-          - name: 'Upload Wheels'
-            uses: actions/upload-artifact@v3
-            with:
-                name: dist_${{ matrix.python-version }}_${{ matrix.platform }}
-                path: wheelhouse/*.whl
-                retention-days: 5
+ 
  upload:
     needs: [build_windows,build_linux_compat,build_linux_fast]
     strategy:
