+++ asset_asrl/test/test_AutoScaling/__init__.py
@@ -1,0 +1,3 @@
+import unittest
+if __name__ == '__main__':
+    unittest.main()
