+++ doc/tutorials/tutorials.rst
@@ -16,6 +16,7 @@ New readers should begin at the beginning, since it's important to understand th
     OptimalControlProblem
     PSIOPT
     AutoMeshGuide
+    NewInterface
     OptimalControlUtilites
