--- featuredev/MultiPhaseCannon.py
@@ -1,248 +1,0 @@
-import numpy as np
-import asset_asrl as ast
-import matplotlib.pyplot as plt
-
-vf        = ast.VectorFunctions
-oc        = ast.OptimalControl
-Args      = vf.Arguments
-
-'''
-This example was taken from the Dymos Optimal control library. It is an excellent
-teaching example of what ODE parameters are used for.
-https://openmdao.github.io/dymos/examples/multi_phase_cannonball/multi_phase_cannonball.html
-Only difference is that we use a simple exponenital atnospheric density function
-
-Assuming constant materiel density, and launch energy we need to find
-the radius of the cannon ball that maximizes range.
-'''
-
-##########################################
-
-g0      =  9.81 
-Lstar   =  1000           ## m
-Tstar   =  60.0           ## sec
-Mstar   =  10             ## kgs
-
-Lstar   =  1           ## m
-Tstar   =  1.0           ## sec
-Mstar   =  1             ## kgs
-
-
-lstar   =  1000           ## m
-tstar   =  60.0           ## sec
-mstar   =  10             ## kgs
-
-
-Astar   =  Lstar/Tstar**2
-Vstar   =  Lstar/Tstar
-Rhostar =  Mstar/Lstar**3
-Estar   =  Mstar*(Vstar**2)
-
-
-CD      = .5
-RhoAir  = 1.225     /Rhostar
-RhoIron = 7870      /Rhostar
-h_scale = 8.44e3    /Lstar
-E0      = 400000    /Estar
-g       = g0/Astar
-
-###########################################
-
-
-def MFunc(rad,RhoIron):return (4/3)*(np.pi*RhoIron)*(rad**3)
-def SFunc(rad):  return np.pi*(rad**2)
-    
-    
-##############################################################################
-
-class Cannon(oc.ODEBase):
-    def __init__(self, CD,RhoAir,RhoIron,h_scale,g):
-        ############################################################
-        args  = oc.ODEArguments(4,0,1)
-        
-        v     = args.XVar(0)
-        gamma = args.XVar(1)
-        h     = args.XVar(2)
-        r     = args.XVar(3)
-        
-        rad = args.PVar(0)
-        
-        S    = SFunc(rad)
-        M    = MFunc(rad,RhoIron)
-        
-        rho     = RhoAir * vf.exp(-h / h_scale)
-        
-        D       = (0.5*CD)*rho*(v**2)*S
-        
-        vdot     = -D/M - g*vf.sin(gamma)
-        gammadot = -g*vf.cos(gamma)/v
-        hdot     = v*vf.sin(gamma)
-        rdot     = v*vf.cos(gamma)
-        
-        ode = vf.stack([vdot,gammadot,hdot,rdot])
-        ##############################################################
-        super().__init__(ode,4,0,1)
-        
-        
-def EFunc():
-    v,rad =  Args(2).tolist()
-    M = MFunc(rad,RhoIron)
-    E = 0.5*M*(v**2)
-    return E - E0
-##############################################################################        
-def Plot(Ascent,Descent):
-    AT = np.array(Ascent).T
-    DT = np.array(Descent).T
-
-    AT[0]*=Vstar
-    DT[0]*=Vstar
-
-    AT[1]*=180/np.pi
-    DT[1]*=180/np.pi
-
-    AT[2]*=Lstar
-    DT[2]*=Lstar
-
-    AT[3]*=Lstar
-    DT[3]*=Lstar
-
-    AT[4]*=Tstar
-    DT[4]*=Tstar
-
-
-    fig = plt.figure()
-    ax0 = plt.subplot(421)
-    ax1 = plt.subplot(423)
-    ax2 = plt.subplot(425)
-    ax3 = plt.subplot(427)
-    axs =[ax0,ax1,ax2,ax3]
-
-    labs = [r'$v\; \frac{m}{s}$',r'$\gamma\;^\circ$', r'Altitude $m$', r' Range $m$']
-
-    for i in range(0,4):
-        
-        axs[i].plot(AT[4],AT[i],color='r')
-        axs[i].plot(DT[4],DT[i],color='b')
-        axs[i].set_ylabel(labs[i])
-        axs[i].set_xlabel(r"$t$")
-        
-        axs[i].grid(True)
-
-
-    ax4 = plt.subplot(122)
-    ax4.plot(AT[3],AT[2],color='r',label='Ascent')
-    ax4.plot(DT[3],DT[2],color='b',label='Descent')
-    ax4.grid(True)
-
-
-    ax4.set_ylabel(" Altitude (m) ")
-    ax4.set_xlabel(" Down Range (m) ")
-    ax4.legend()
-    
-    fig.set_size_inches(15.0, 7.5, forward=True)
-    fig.tight_layout()
-
-    plt.show()
-        
-
-##############################################################################        
-if __name__ == "__main__":
-
-    rad0   = .1 /Lstar
-    h0     = 100 /Lstar
-    r0     = 0
-    m0     = MFunc(rad0,RhoIron)
-    gamma0 = np.deg2rad(45)
-    v0     = np.sqrt(2*E0/m0)*.99
-    
-    
-    
-    ode = Cannon(CD,RhoAir,RhoIron,h_scale,g)
-    integ = ode.integrator(.01)
-    integ.setAbsTol(1.0e-12)    
-    
-    IG = np.zeros((6))
-    IG[0] = v0
-    IG[1] =gamma0
-    IG[2] = h0
-    IG[3] = r0
-    IG[5] = rad0
-    
-    
-    def AscentEvent():
-        args  = oc.ODEArguments(4,0,1)
-        return args[0]*vf.sin(args[1])
-    
-    AscentIG = integ.integrate_dense(IG,60/Tstar,[(AscentEvent(),0,1)])[0]
-
-    DescentIG = integ.integrate_dense(AscentIG[-1],AscentIG[-1][4]+ 1000/Tstar,
-                                      [( oc.ODEArguments(4,0,1)[2],0,1)])[0]
-    ##########################################################################
-    
-    Units = np.ones((6))
-    
-    Units[0]=lstar/tstar
-    Units[2]=lstar
-    Units[3]=lstar
-    Units[4]=tstar
-    Units[5]=lstar
-    
-    tmode = "LGL5"
-    nsegs = 128
-    
-    AutoScaling=True
-    
-    aphase = ode.phase(tmode,AscentIG,nsegs)
-    aphase.AutoScaling=AutoScaling
-    aphase.setUnits(Units)
-    aphase.addLowerVarBound("ODEParams",0,0.0,1,AutoScale='auto')
-    aphase.addLowerVarBound("Front",1,0.0,1.0,AutoScale='auto')
-    aphase.addBoundaryValue("Front",[2,3,4],[h0,r0,0],AutoScale='auto')
-    
-    aphase.addInequalCon("Front",EFunc(),[0],[0],[],AutoScale=None)
-    aphase.addBoundaryValue("Back",[1],[0.0],AutoScale='auto')
-        
-    dphase = ode.phase(tmode,DescentIG,nsegs)
-    dphase.AutoScaling=AutoScaling
-    dphase.setUnits(Units)
-    
-    dphase.addBoundaryValue("Back",[2],[0.0],AutoScale='auto')
-    dphase.addValueObjective("Back",3,-1.0,AutoScale='auto')
-    
-    ocp = oc.OptimalControlProblem()
-    ocp.AutoScaling=AutoScaling
-    ocp.addPhase(aphase)
-    ocp.addPhase(dphase)
-    
-    
-    # Enforce continuatiy in time dependent vars
-    ocp.addForwardLinkEqualCon(aphase,dphase,[0,1,2,3,4])
-    
-    F = Args(2)[0]-Args(2)[1]
-    
-    
-    
-    ocp.addLinkEqualCon(F,aphase,"ODEParams",0,
-                              dphase,"ODEParams",0,AutoScale=1)
-    
-    
-    ocp.optimizer.set_OptLSMode("L1")
-    ocp.optimize()
-    
-    
-    Ascent  = aphase.returnTraj()
-    Descent = dphase.returnTraj()
-    
-    print("Launch Angle:",Ascent[0][1]*180/np.pi," deg")
-    print("Optimized Distance:",Descent[-1][3]*Lstar," m")
-    print("Optimized Radius:"  ,Descent[-1][-1]*Lstar," m")
-
-    Plot(Ascent,Descent)
-
-    ##########################################################################
-
-
-
-
-
-
